Rename the generated-code folder in the server package
