+++ packages/server/src/api/todo/adx/index.ts
@@ -1,4 +1,4 @@
-import { Server } from '../../../xrpc'
+import { Server } from '../../../lexicon'
 import names from './names'
 import session from './session'
 import account from './account'
