+++ packages/server/src/api/todo/social/getPostThread.ts
@@ -1,7 +1,7 @@
-import { Server } from '../../../xrpc'
+import { Server } from '../../../lexicon'
 import { AuthRequiredError } from '@adxp/xrpc-server'
 import { DataSource } from 'typeorm'
-import * as GetPostThread from '../../../xrpc/types/todo/social/getPostThread'
+import * as GetPostThread from '../../../lexicon/types/todo/social/getPostThread'
 import { PostIndex } from '../../../db/records/post'
 import { ProfileIndex } from '../../../db/records/profile'
 import { UserDid } from '../../../db/user-dids'
