+++ packages/server/src/db/index.ts
@@ -1,9 +1,9 @@
-import * as Badge from '../xrpc/types/todo/social/badge'
-import * as Follow from '../xrpc/types/todo/social/follow'
-import * as Like from '../xrpc/types/todo/social/like'
-import * as Post from '../xrpc/types/todo/social/post'
-import * as Profile from '../xrpc/types/todo/social/profile'
-import * as Repost from '../xrpc/types/todo/social/repost'
+import * as Badge from '../lexicon/types/todo/social/badge'
+import * as Follow from '../lexicon/types/todo/social/follow'
+import * as Like from '../lexicon/types/todo/social/like'
+import * as Post from '../lexicon/types/todo/social/post'
+import * as Profile from '../lexicon/types/todo/social/profile'
+import * as Repost from '../lexicon/types/todo/social/repost'
 import { DataSource } from 'typeorm'
 import { ValidationResult, ValidationResultCode } from '@adxp/lexicon'
 import { DbRecordPlugin } from './types'
